Factor out DocumentPermissionsField component
