+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/content_sources/components/add_source/connect_instance.tsx
@@ -11,7 +11,6 @@ import { useActions, useValues } from 'kea';
 
 import {
   EuiButton,
-  EuiCallOut,
   EuiFieldText,
   EuiFlexGroup,
   EuiFlexItem,
@@ -21,31 +20,22 @@ import {
   EuiLink,
   EuiPanel,
   EuiSpacer,
-  EuiSwitch,
   EuiText,
-  EuiTitle,
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { FormattedMessage } from '@kbn/i18n-react';
 
 import { LicensingLogic } from '../../../../../shared/licensing';
 import { AppLogic } from '../../../../app_logic';
 import { EXPLORE_PLATINUM_FEATURES_LINK } from '../../../../constants';
-import { DOCUMENT_PERMISSIONS_DOCS_URL, ENT_SEARCH_LICENSE_MANAGEMENT } from '../../../../routes';
+import { ENT_SEARCH_LICENSE_MANAGEMENT } from '../../../../routes';
 import { FeatureIds, Configuration, Features } from '../../../../types';
-import { LEARN_MORE_LINK } from '../../constants';
 
 import { AddSourceLogic } from './add_source_logic';
 import {
-  CONNECT_WHICH_OPTION_LINK,
-  CONNECT_DOC_PERMISSIONS_LABEL,
-  CONNECT_DOC_PERMISSIONS_TITLE,
-  CONNECT_NEEDS_PERMISSIONS,
-  CONNECT_NOT_SYNCED_TITLE,
-  CONNECT_NOT_SYNCED_TEXT,
   SOURCE_FEATURES_DOCUMENT_LEVEL_PERMISSIONS_FEATURE,
   SOURCE_FEATURES_DOCUMENT_LEVEL_PERMISSIONS_TITLE,
 } from './constants';
+import { DocumentPermissionsField } from './document_permissions_field';
 import { SourceFeatures } from './source_features';
 
 interface ConnectInstanceProps {
@@ -147,66 +137,12 @@ export const ConnectInstance: React.FC<ConnectInstanceProps> = ({
     </>
   );
 
-  const whichDocsLink = (
-    <EuiLink target="_blank" href={DOCUMENT_PERMISSIONS_DOCS_URL}>
-      {CONNECT_WHICH_OPTION_LINK}
-    </EuiLink>
-  );
-
   const permissionField = (
-    <>
-      <EuiPanel paddingSize="l" hasShadow={false} color="subdued">
-        <EuiTitle size="s">
-          <h1>
-            <strong>{CONNECT_DOC_PERMISSIONS_TITLE}</strong>
-          </h1>
-        </EuiTitle>
-        <EuiSpacer size="s" />
-        <EuiText color="subdued" size="s">
-          {!needsPermissions && (
-            <span>
-              <FormattedMessage
-                id="xpack.enterpriseSearch.workplaceSearch.contentSource.connect.docPermissionsUnavailable.message"
-                defaultMessage="Document-level permissions are not yet available for this source. {link}"
-                values={{
-                  link: (
-                    <EuiLink target="_blank" href={DOCUMENT_PERMISSIONS_DOCS_URL}>
-                      {LEARN_MORE_LINK}
-                    </EuiLink>
-                  ),
-                }}
-              />
-            </span>
-          )}
-          {needsPermissions && indexPermissionsValue && (
-            <span>
-              {CONNECT_NEEDS_PERMISSIONS}
-              <EuiSpacer size="s" />
-              {whichDocsLink}
-            </span>
-          )}
-        </EuiText>
-        {!indexPermissionsValue && (
-          <>
-            <EuiCallOut title={CONNECT_NOT_SYNCED_TITLE} color="warning">
-              <p>
-                {CONNECT_NOT_SYNCED_TEXT}
-                {needsPermissions && whichDocsLink}
-              </p>
-            </EuiCallOut>
-          </>
-        )}
-        <EuiSpacer />
-        <EuiSwitch
-          label={<strong>{CONNECT_DOC_PERMISSIONS_LABEL}</strong>}
-          name="index_permissions"
-          onChange={(e) => setSourceIndexPermissionsValue(e.target.checked)}
-          checked={indexPermissionsValue}
-          disabled={!needsPermissions}
-        />
-      </EuiPanel>
-      <EuiSpacer size="xl" />
-    </>
+    <DocumentPermissionsField
+      needsPermissions={needsPermissions}
+      indexPermissionsValue={indexPermissionsValue}
+      setValue={setSourceIndexPermissionsValue}
+    />
   );
 
   const documentLevelPermissionsCallout = (
